Fix bug in django.contrib.gis.geoip2.GeoIP2 initialisation
There is a bug in Django's implementation here - https://github.com/django/django/blob/1.11.9/django/contrib/gis/geoip2/base.py#L117

Basically, if a GEOIP_PATH is supplied, but there is no database file
at that location, then during initialisation neither `_country` nor `_city`
attributes are set, which means that the `GeoIP2._reader` attr will
return None. This is a problem as a number of methods, including the
`__repr__` method assume that this will always be set to something.